example-based: split counterfactuals and semifactuals
